Refactor ExerciseList
