+++ src/main/java/seedu/fitnus/user/User.java
@@ -2,21 +2,19 @@ package seedu.fitnus.user;
 
 import seedu.fitnus.Date;
 import seedu.fitnus.Drink;
-import seedu.fitnus.Exercise;
-import seedu.fitnus.ExerciseIntensity;
+import seedu.fitnus.exercise.Exercise;
 import seedu.fitnus.Meal;
+import seedu.fitnus.exercise.ExerciseList;
 import seedu.fitnus.parser.Parser;
 import seedu.fitnus.Water;
 
 import seedu.fitnus.exception.IncompleteDeleteException;
 import seedu.fitnus.exception.IncompleteDrinkException;
 import seedu.fitnus.exception.IncompleteEditException;
-import seedu.fitnus.exception.IncompleteExerciseException;
 import seedu.fitnus.exception.IncompleteMealException;
 import seedu.fitnus.exception.InvalidListIndexException;
 import seedu.fitnus.exception.NegativeValueException;
 import seedu.fitnus.exception.UnregisteredDrinkException;
-import seedu.fitnus.exception.UnregisteredExerciseException;
 import seedu.fitnus.exception.UnregisteredMealException;
 import seedu.fitnus.exception.InvalidDateException;
 
@@ -32,22 +30,21 @@ public class User {
     public static ArrayList<Meal> mealList;
     public static ArrayList<Drink> drinkList;
     public static ArrayList<Water> waterList;
-    public static ArrayList<Exercise> exerciseList;
     // list for all dates except today
     public static ArrayList<Meal> mealListAll;
     public static ArrayList<Drink> drinkListAll;
     public static ArrayList<Water> waterListAll;
-    public static ArrayList<Exercise> exerciseListAll;
+    public static ExerciseList myExerciseList;
 
     public User() {
         mealList = new ArrayList<>();
         drinkList = new ArrayList<>();
-        exerciseList = new ArrayList<>();
         waterList = new ArrayList<>();
         mealListAll = new ArrayList<>();
         drinkListAll = new ArrayList<>();
-        exerciseListAll = new ArrayList<>();
         waterListAll = new ArrayList<>();
+
+        myExerciseList = new ExerciseList();
     }
 
     public void handleAddNewMealNutrient(String command) throws NegativeValueException{
@@ -75,17 +72,6 @@ public class User {
         Drink.nutrientDetails.put(description, new int[]{calories, carbs, sugar, protein, fat});
 
         System.out.println("Added " + description + " to available drinks");
-    }
-
-    public void handleAddNewExerciseCalories(String command) throws NegativeValueException{
-        Parser.parseNewExercise(command);
-        String description = Parser.exerciseCaloriesDescription;
-        int high = Parser.exerciseCaloriesHigh;
-        int medium = Parser.exerciseCaloriesMedium;
-        int low = Parser.exerciseCaloriesLow;
-        Exercise.exerciseDetails.put(description, new int[]{high, medium, low});
-
-        System.out.println("Added " + description + " to available exercises");
     }
 
     /**
@@ -155,7 +141,7 @@ public class User {
         for (Drink drink: drinkList) {
             caloriesCount += drink.getCalories();
         }
-        for (Exercise exercise: exerciseList) {
+        for (Exercise exercise: myExerciseList.exerciseList) {
             caloriesCount -= exercise.getCaloriesBurnt();
         }
         System.out.println("Total Calories: " + caloriesCount);
@@ -256,21 +242,6 @@ public class User {
             Meal currentMeal = mealListToPrint.get(i);
             System.out.println((startIndex+i) + ". " + currentMeal.getName() + " (serving size: "
                     + currentMeal.getServingSize() + ")" + " | date: " + currentMeal.getDate());
-        }
-    }
-
-    /**
-     * Prints all the exercises in the exerciseListToPrint,
-     * inclusive of the duration, intensity and date.
-     *
-     * @param exerciseListToPrint arraylist containing the exercises that should be printed
-     */
-    public void printExerciseList(ArrayList<Exercise> exerciseListToPrint) {
-        for (int i = 0; i < exerciseListToPrint.size(); i++) {
-            Exercise currentExercise = exerciseListToPrint.get(i);
-            System.out.println((i+1) + ". " + currentExercise.getName() + " | duration: " +
-                    currentExercise.getDuration() + " | intensity: " + currentExercise.getIntensity()
-                    + " | date: " + currentExercise.getDate());
         }
     }
 
@@ -420,65 +391,6 @@ public class User {
     }
 
     /**
-     * Handles when the user is listing the exercises they have done today.
-     * Method first checks if the list is empty.
-     */
-    public void handleListExercises() {
-        System.out.println("here's the exercises you've done today");
-        if (exerciseList.isEmpty()) {
-            System.out.println("  >> nothing so far :o");
-        } else {
-            printExerciseList(exerciseList);
-        }
-    }
-
-    /**
-     * Handles when the user is listing all exercises they have done, inclusive of previously saved exercises.
-     * Method first checks if the list is empty.
-     */
-
-    public void handleListExercisesAll() {
-        System.out.println("here's the exercises you've done so far");
-        if (exerciseListAll.isEmpty() && exerciseList.isEmpty()) {
-            System.out.println("  >> nothing so far :o");
-        } else {
-            ArrayList<Exercise> appendedExerciseList = new ArrayList<>();
-            appendedExerciseList.addAll(exerciseListAll);
-            appendedExerciseList.addAll(exerciseList);
-            printExerciseList(appendedExerciseList);
-        }
-    }
-
-    /**
-     * Handles when the user is listing the exercises they have done on a certain date.
-     * Method will first extract all exercises that have this corresponding date,
-     * before printing.
-     *
-     * @param command string inputted by the user, containing the date of which they would like to list exercises of
-     * @throws InvalidDateException if the date inputted by user is invalid
-     */
-    public void handleListExercisesDate(String command) throws InvalidDateException {
-        String date = Parser.parseListDate(command);
-        ArrayList<Exercise> exercisesListDate = new ArrayList<>();
-        for (Exercise e : exerciseListAll) {
-            if (e.getDate().equals(date)) {
-                exercisesListDate.add(e);
-            }
-        }
-        for (Exercise e : exerciseList) {
-            if (e.getDate().equals(date)) {
-                exercisesListDate.add(e);
-            }
-        }
-        System.out.println("here's the exercises you've done on " + date);
-        if (exercisesListDate.isEmpty()) {
-            System.out.println("  >> nothing so far :o");
-        } else {
-            printExerciseList(exercisesListDate);
-        }
-    }
-
-    /**
      * Handles when the user is listing all meals and drinks they have inputted today.
      * Method first checks if the lists is empty.
      */
@@ -496,7 +408,7 @@ public class User {
         }
 
         System.out.println("       ~~~");
-        handleListExercises();
+        myExerciseList.handleListExercises();
     }
 
     /**
@@ -520,7 +432,7 @@ public class User {
         }
 
         System.out.println("       ~~~");
-        handleListExercisesAll();
+        myExerciseList.handleListExercisesAll();
     }
 
     /**
@@ -568,7 +480,7 @@ public class User {
         }
 
         System.out.println("       ~~~");
-        handleListExercisesDate(command);
+        myExerciseList.handleListExercisesDate(command);
     }
 
     /**
@@ -679,50 +591,6 @@ public class User {
         System.out.println("Removed " + drinkName + " from drinks");
     }
 
-    /**
-     * Handles when the user would like to delete a previously inputted exercise.
-     *
-     * @param command string inputted by the user, containing the index of the exercise to delete
-     * @throws InvalidListIndexException if the provided index is not a valid index in exerciseList
-     * @throws IncompleteDeleteException if the user did not comply with the required command format
-     */
-    public void handleDeleteExercise(String command) throws InvalidListIndexException, IncompleteDeleteException {
-        if (command.length() < 16) {
-            throw new IncompleteDeleteException();
-        }
-
-        int exerciseIndex = Integer.parseInt(command.substring(15).trim()) - 1;
-        if (exerciseIndex >= exerciseList.size() || exerciseIndex < 0) {
-            throw new InvalidListIndexException();
-        }
-
-        String exerciseName = exerciseList.get(exerciseIndex).getName();
-        exerciseList.remove(exerciseIndex);
-        System.out.println("Removed " + exerciseName + " from exercises done");
-    }
-
-    /**
-     * Adds an exercise to the user's current exerciseList, based on what exercise the user has done,
-     * its duration and intensity.
-     *
-     * @param command string inputted by the user, containing the exercise they have done, its duration and
-     *                intensity.
-     * @throws IncompleteExerciseException if the user did not comply with the required format
-     * @throws UnregisteredExerciseException if the user has inputted an exercise that was not pre-defined
-     * @throws NegativeValueException if the provided exercise duration is a negative value
-     */
-    public void handleExercise(String command) throws IncompleteExerciseException, UnregisteredExerciseException,
-            NegativeValueException {
-        Parser.parseExercise(command);
-        String exerciseType = Parser.exerciseDescription;
-        int duration = Parser.exerciseDuration;
-        ExerciseIntensity intensity = Parser.exerciseIntensity;
-        Date currentDate = new Date();
-        exerciseList.add(new Exercise(exerciseType, duration, intensity, currentDate.getDate()));
-        assert !exerciseList.isEmpty(): "failed to track exercise";
-
-        System.out.println("Tracked " + duration + " minutes of " + exerciseType);
-    }
 
     /**
      * Handle when user would like to clear all entries from today.
@@ -732,24 +600,13 @@ public class User {
         mealList.clear();
         drinkList.clear();
         waterList.clear();
-        exerciseList.clear();
+        myExerciseList.exerciseList.clear();
 
         assert mealList.isEmpty(): "clearing of meal list failed";
         assert drinkList.isEmpty(): "clearing of drink list failed";
-        assert exerciseList.isEmpty(): "clearing of exercise list failed";
+        assert myExerciseList.exerciseList.isEmpty(): "clearing of exercise list failed";
 
         System.out.println("All entries have been deleted");
-    }
-
-    /**
-     * Prints the number of calories the user has burnt today.
-     */
-    public void handleCaloriesBurnt() {
-        int caloriesBurnt = 0;
-        for (Exercise exercise: exerciseList) {
-            caloriesBurnt += exercise.getCaloriesBurnt();
-        }
-        System.out.println("Total calories burnt: " + caloriesBurnt);
     }
 
     /**
@@ -776,7 +633,7 @@ public class User {
         for (Drink drink: drinkList) {
             caloriesCount += drink.getCalories();
         }
-        for (Exercise exercise: exerciseList) {
+        for (Exercise exercise: myExerciseList.exerciseList) {
             caloriesCount -= exercise.getCaloriesBurnt();
         }
         int caloriesDifference = RECOMMEND_CALORIE_INTAKE - caloriesCount;
